requirements.txt
boto3
chromadb
torch
sentence-transformers
pytesseract
Pillow
pymupdf
fastapi
uvicorn
boto3
pydantic
python-dotenv
chroma
jinja2
scikit-learn
python-multipart
requests
docx
python-docx
scipy
numpy
uuid4
PyPDF2